Clarify $http import
When I first tried to work with this, I was at a loss as to why the code wasn't running as the Nuxt HTTP module wasn't properly configured to work in my project. So, I have explained where this comes from and how to add the module to prevent other people reading the docs from having the same issue.